Testing...

+++ fft/fft.c
@@ -3,9 +3,9 @@
 #include "fft.h"
 
 /*exp (-i*2*Pi*(k/N)*n)*/
-c_vector * fftv(c_vector * r, const c_vector * x, const unsigned int N, const unsigned int k)
+c_vector * fftv(c_vector * r, const c_vector * x, const unsigned long int N, const unsigned long int k)
 {
-  unsigned int n;
+  unsigned long int n;
   c_vector f,ret;
 
   ret.f[REAL] = 0;
@@ -25,9 +25,9 @@ c_vector * fftv(c_vector * r, const c_vector * x, const unsigned int N, const un
 }
 
 
-complex * fft(complex * r, const complex * x, const unsigned int N, const unsigned int k)
+complex * fft(complex * r, const complex * x, const unsigned long int N, const unsigned long int k)
 {
-  unsigned int n;
+  unsigned long int n;
   complex f,ret;
 
   ret.real = 0;
