+++ fft/fft.h
@@ -1,3 +1,2 @@
-c_vector * fftv(c_vector * r, const c_vector * x, const unsigned int N, const unsigned int k);
-complex * fft(complex * r, const complex * x, const unsigned int N, const unsigned int k);
-
+c_vector * fftv(c_vector * r, const c_vector * x, const unsigned long int N, const unsigned long int k);
+complex * fft(complex * r, const complex * x, const unsigned long int N, const unsigned long int k);
